Crate pallet_bridge_relayers

Runtime module that is used to store relayer rewards and (in the future) to coordinate relations between relayers.

Re-exports

pub use weights::WeightInfo;

pub use pallet::*;

Modules

extension

Signed extension, built around pallet-bridge-relayers. It is able to:

migration

A module that is responsible for migration of storage.

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

weights

Autogenerated weights for pallet_bridge_relayers

Structs

DeliveryConfirmationPaymentsAdapter

Adapter that allows relayers pallet to be used as a delivery+dispatch payment mechanism for the pallet-bridge-messages pallet and using RewardsAccountParams.

PayRewardFromAccount

Reward payment procedure that executes a balances::transfer call from the account derived from the given RewardsAccountParams to the relayer or an alternative beneficiary.

StakeAndSlashNamed

StakeAndSlash that works with NamedReservableCurrency and uses named reservations.

Constants

LOG_TARGET

The target that will be used when publishing logs related to this pallet.

Traits

RewardLedger

A trait defining a reward ledger, which tracks rewards that can be later claimed.

WeightInfoExt

Extended weight info.