Trait pallet_bridge_relayers::pallet::Config

pub trait Config: Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type Reward: AtLeast32BitUnsigned + Copy + Parameter + MaxEncodedLen;
    type PaymentProcedure: PaymentProcedure<Self::AccountId, Self::Reward>;
    type StakeAndSlash: StakeAndSlash<Self::AccountId, BlockNumberFor<Self>, Self::Reward>;
    type WeightInfo: WeightInfoExt;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

The overarching event type.

type Reward: AtLeast32BitUnsigned + Copy + Parameter + MaxEncodedLen

Type of relayer reward.

type PaymentProcedure: PaymentProcedure<Self::AccountId, Self::Reward>

Pay rewards scheme.

type StakeAndSlash: StakeAndSlash<Self::AccountId, BlockNumberFor<Self>, Self::Reward>

Stake and slash scheme.

type WeightInfo: WeightInfoExt

Pallet call weights.

Object Safety

This trait is not object safe.