Type Alias pallet_bridge_relayers::pallet::RegisteredRelayers

pub type RegisteredRelayers<T: Config> = StorageMap<_GeneratedPrefixForStorageRegisteredRelayers<T>, Blake2_128Concat, T::AccountId, Registration<BlockNumberFor<T>, T::Reward>, OptionQuery>;

Relayers that have reserved some of their balance to get free priority boost for their message delivery transactions.

Other relayers may submit transactions as well, but they will have default priority and will be rejected (without significant tip) in case if registered relayer is present.

Storage type is [StorageMap] with key type T :: AccountId and value type Registration < BlockNumberFor < T > , T :: Reward >.

Aliased Type

struct RegisteredRelayers<T: Config>(/* private fields */);