Module pallet_bridge_relayers::extension

Signed extension, built around pallet-bridge-relayers. It is able to:

refund the cost of successful message delivery and confirmation transactions to the submitter by registering corresponding reward in the pallet;

bump priority of messages delivery and confirmation transactions, signed by the registered relayers.

Modules

per_message

Computations, specific to bridge messages transactions.

per_parachain_header

Computations, specific to bridge parachains transactions.

per_relay_header

Computations, specific to bridge relay chains transactions.

Structs

BridgeRelayersSignedExtension

A signed extension, built around pallet-bridge-relayers.

PreDispatchData

Data that is crafted in pre_dispatch method and used at post_dispatch.

WithGrandpaChainExtensionConfig

Adapter to be used in signed extension configuration, when bridging with remote chains that are using GRANDPA finality.

WithMessagesExtensionConfig

Transaction extension that refunds a relayer for standalone messages delivery and confirmation transactions. Finality transactions are not refunded.

WithParachainExtensionConfig

Adapter to be used in signed extension configuration, when bridging with remote parachains.

Enums

RelayerAccountAction

The actions on relayer account that need to be performed because of his actions.

Functions

compute_priority_boost

Compute priority boost for transaction that brings given number of bridge items (messages, headers, …), when every additional item adds PriorityBoostPerItem to transaction priority.

Type Aliases

ItemCount

We’ll deal with different bridge items here - messages, headers, … To avoid being too verbose with generic code, let’s just define a separate alias.