Trait pallet_bridge_relayers::WeightInfoExt

pub trait WeightInfoExt: WeightInfo {
    // Provided methods
    fn receive_messages_proof_overhead_from_runtime() -> Weight { ... }
    fn receive_messages_delivery_proof_overhead_from_runtime() -> Weight { ... }
    fn extra_weight_of_successful_receive_messages_proof_call() -> Weight { ... }
}

Extended weight info.

Provided Methods

fn receive_messages_proof_overhead_from_runtime() -> Weight

Returns weight, that needs to be added to the pre-dispatch weight of message delivery call, if RefundBridgedParachainMessages signed extension is deployed at runtime level.

fn receive_messages_delivery_proof_overhead_from_runtime() -> Weight

Returns weight, that needs to be added to the pre-dispatch weight of message delivery confirmation call, if RefundBridgedParachainMessages signed extension is deployed at runtime level.

fn extra_weight_of_successful_receive_messages_proof_call() -> Weight

Returns weight that we need to deduct from the message delivery call weight that has completed successfully.

Usually, the weight of slash_and_deregister is larger than the weight of the register_relayer_reward. So if relayer has been rewarded, we want to deduct the difference to get the actual post-dispatch weight.

Object Safety

This trait is not object safe.